+++ test/fixtures3/modcache/github.com/acme/worker@v1.0.0/worker.go
@@ -1,0 +1,5 @@
+package worker
+
+type ExplicitDependencyWorker struct{}
+
+func (*ExplicitDependencyWorker) ResolveExplicitDependency(string) error { return nil }