+++ test/fixtures3/proj/service.go
@@ -1,0 +1,5 @@
+package impl
+
+type ExplicitDependencyService interface {
+	ResolveExplicitDependency(string) error
+}